requirements.txt
asgiref==3.8.1
Django==4.2.14
djangorestframework==3.15.2
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
